+++ CHANGES.md
@@ -3,3 +3,4 @@
 
 # 0.0.2 (2018-01-22)
 - Modify default config to match MQ Docker container image defaults
+- add TLS related properties